requirements.txt
customtkinter>=5.2.0
hid>=1.0.5
pyobjc-framework-Quartz>=10.0
pynput>=1.7.6
Pillow>=10.0.0
